:truck: Rename types file
