+++ src/app/_types/Card.ts
